ipq806x: restore recent changes made to 6.6 dts files
911fd4ab3be4 (2025-10-14) restored dts files for kernel 6.6 to files-6.6
after the preceding 90c668cac489 moved files-6.6 to files-6.12 to be
used by the new testing kernel, 6.12. This restoration omitted
5c210483e62e (2025-10-08), which fixed important aspects of the wifi
device definitions in qcom-ipq8064-eax500.dtsi and
qcom-ipq8064-unifi-ac-hd.dts.

The 5c210483e62e fix persisted into the 6.12 dts files. This change now
restores it for 6.6. Note that ipq806x is currently using kernel 6.6 for
most purposes, as 6.12 is only in testing status for this target.

This change was prepared with `git diff
911fd4ab3be4:target/linux/ipq806x/files-6.6
90c668cac489^:target/linux/ipq806x/files-6.6`. This identified
0b5b7bd716c0 (2025-10-08) as another change that was not present in the
restored files-6.6, but it has since been superseded by 475ce695751a
(2025-10-20), so no action is necessary to restore it. There were no
diffs present between these revisions in config-6.6 or patches-6.6,
which were also moved and restored in the 6.12 bring-up.

Signed-off-by: Mark Mentovai <mark@mentovai.com>
Cc: Shiji Yang <yangshiji66@outlook.com>
Cc: Rosen Penev <rosenp@gmail.com>
Link: https://github.com/openwrt/openwrt/pull/20464
Signed-off-by: Robert Marko <robimarko@gmail.com>